migrate struct to object
